+++ packages/table/src/filters/DateFilter/declarations.ts
@@ -1,0 +1,12 @@
+export type TDateFilterValue = {
+  value: number | Date;
+  secondValue?: number | Date;
+  operator:
+    | 'equals'
+    | 'notEquals'
+    | 'greater'
+    | 'greaterOrEqual'
+    | 'less'
+    | 'lessOrEqual'
+    | 'between';
+};