+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [13.4.2](https://github.com/DevoInc/genesys-ui/compare/v13.4.1...v13.4.2) (2024-10-29)
+
+
+### Bug Fixes
+
+* **table:** sortable only header cell ([5e02e17](https://github.com/DevoInc/genesys-ui/commit/5e02e17bfa7dd66912cda2ed7158bb5a33197b55))
+
+## [13.4.1](https://github.com/DevoInc/genesys-ui/compare/v13.4.0...v13.4.1) (2024-10-29)
+
+## [13.4.0](https://github.com/DevoInc/genesys-ui/compare/v13.3.3...v13.4.0) (2024-10-28)
+
+
+### Features
+
+* added new collapsable variant for FormGroup ([5d02410](https://github.com/DevoInc/genesys-ui/commit/5d02410f9267efd4192c7f4baf69cd65a77ed183))
+
 ## [13.3.3](https://github.com/DevoInc/genesys-ui/compare/v13.3.2...v13.3.3) (2024-10-24)